requirements.txt
python-dotenv==1.0.1
fastapi[standard]==0.115.3
starlette==0.41.2
starlette-session==0.4.3
psycopg==3.2.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.16
authlib==1.3.2
itsdangerous==2.2.0
pycoingecko==3.2.0
pandas==2.2.3
pymongo==3.11